# seedbox-service/openapi.yaml
openapi: 3.0.3
info:
  title: qBittorrent WebUI API
  description: |
    - All API methods are under `/api/v2/APIName/methodName`, where `APIName` is a certain subgroup of API methods whose functionality is related.
    - Either `GET` or `POST` can be used as the request type for all API methods.
    - All API methods require [authentication](https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#authentication) (except the `/api/v2/auth/login` method itself, obviously).
  termsOfService: 'https://github.com/qbittorrent/qBittorrent/blob/master/COPYING'
  license:
    name: GNU GPLv2
    url: 'https://www.gnu.org/licenses/old-licenses/gpl-2.0.en.html'
  version: 2.8.3
servers:
  - url: '/api/v2'
    description: qBittorrent Web UI API
paths:
  /auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: |
        Upon success, the response will contain a cookie with your SID. You must supply the cookie whenever you want to perform an operation that requires authentication.
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#login'
        description: WebUI-API on qBittorrent wiki
      operationId: authLoginPost
      parameters:
        - in: header
          name: Referer
          schema:
            type: string
        - in: header
          name: Origin
          schema:
            type: string
      requestBody:
        description: |
          Example showing how to login and execute a command that requires authentication using `curl`:

          ```
          $ curl -i --header 'Referer: http://localhost:8080' --data 'username=admin&password=adminadmin' http://localhost:8080/api/v2/auth/login
          HTTP/1.1 200 OK
          Content-Encoding:
          Content-Length: 3
          Content-Type: text/plain; charset=UTF-8
          Set-Cookie: SID=hBc7TxF76ERhvIw0jQQ4LZ7Z1jQUV0tQ; path=/
          $ curl http://localhost:8080/api/v2/torrents/info --cookie "SID=hBc7TxF76ERhvIw0jQQ4LZ7Z1jQUV0tQ"
          ```

          Note: Set `Referer` or `Origin` header to the exact same domain and port as used in the HTTP query `Host` header.
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
              required:
                - username
                - password
              example: username=admin&password=adminadmin
        required: true
      responses:
        '200':
          description: All other scenarios
          headers:
            Set-Cookie:
              schema:
                type: string
                example: SID=hBc7TxF76ERhvIw0jQQ4LZ7Z1jQUV0tQ
          content:
            text/plain:
              schema:
                type: string
        '403':
          description: User's IP is banned for too many failed login attempts
...
  /app/preferences:
    get:
      tags:
        - app
      summary: Get application preferences
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-application-preferences'
        description: WebUI-API on qBittorrent wiki
      operationId: appPreferencesGet
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
                description: The response is a JSON object with several fields (key-value) pairs representing the application's settings. The contents may vary depending on which settings are present in qBittorrent.ini.
                example: |
                  {
                    "add_trackers": "",
                    "add_trackers_enabled": false,
                    "alt_dl_limit": 10240,
                    "alt_up_limit": 10240,
                    "alternative_webui_enabled": false,
                    "alternative_webui_path": "/home/user/Documents/qbit-webui",
                    "announce_ip": "",
                    "announce_to_all_tiers": true,
                    "announce_to_all_trackers": false,
                    "anonymous_mode": false,
                    "async_io_threads": 4,
                    "auto_delete_mode": 0,
                    "auto_tmm_enabled": false,
                    "autorun_enabled": false,
                    "autorun_program": "",
                    "banned_IPs": "",
                    "bittorrent_protocol": 0,
                    "bypass_auth_subnet_whitelist": "",
                    "bypass_auth_subnet_whitelist_enabled": false,
                    "bypass_local_auth": false,
                    "category_changed_tmm_enabled": false,
                    "checking_memory_use": 32,
                    "create_subfolder_enabled": true,
                    "current_interface_address": "",
                    "current_network_interface": "",
                    "dht": true,
                    "disk_cache": -1,
                    "disk_cache_ttl": 60,
                    "dl_limit": 0,
                    "dont_count_slow_torrents": false,
                    "dyndns_domain": "changeme.dyndns.org",
                    "dyndns_enabled": false,
                    "dyndns_password": "",
                    "dyndns_service": 0,
                    "dyndns_username": "",
                    "embedded_tracker_port": 9000,
                    "enable_coalesce_read_write": false,
                    "enable_embedded_tracker": false,
                    "enable_multi_connections_from_same_ip": false,
                    "enable_os_cache": true,
                    "enable_piece_extent_affinity": false,
                    "enable_upload_suggestions": false,
                    "encryption": 0,
                    "export_dir": "/home/user/Downloads/all",
                    "export_dir_fin": "/home/user/Downloads/completed",
                    "file_pool_size": 40,
                    "incomplete_files_ext": false,
                    "ip_filter_enabled": false,
                    "ip_filter_path": "",
                    "ip_filter_trackers": false,
                    "limit_lan_peers": true,
                    "limit_tcp_overhead": false,
                    "limit_utp_rate": true,
                    "listen_port": 58925,
                    "locale": "en",
                    "lsd": true,
                    "mail_notification_auth_enabled": false,
                    "mail_notification_email": "",
                    "mail_notification_enabled": false,
                    "mail_notification_password": "",
                    "mail_notification_sender": "qBittorrent_notification@example.com",
                    "mail_notification_smtp": "smtp.changeme.com",
                    "mail_notification_ssl_enabled": false,
                    "mail_notification_username": "",
                    "max_active_downloads": 3,
                    "max_active_torrents": 5,
                    "max_active_uploads": 3,
                    "max_connec": 500,
                    "max_connec_per_torrent": 100,
                    "max_ratio": -1,
                    "max_ratio_act": 0,
                    "max_ratio_enabled": false,
                    "max_seeding_time": -1,
                    "max_seeding_time_enabled": false,
                    "max_uploads": -1,
                    "max_uploads_per_torrent": -1,
                    "outgoing_ports_max": 0,
                    "outgoing_ports_min": 0,
                    "pex": true,
                    "preallocate_all": false,
                    "proxy_auth_enabled": false,
                    "proxy_ip": "0.0.0.0",
                    "proxy_password": "",
                    "proxy_peer_connections": false,
                    "proxy_port": 8080,
                    "proxy_torrents_only": false,
                    "proxy_type": 0,
                    "proxy_username": "",
                    "queueing_enabled": false,
                    "random_port": false,
                    "recheck_completed_torrents": false,
                    "resolve_peer_countries": true,
                    "rss_auto_downloading_enabled":true,
                    "rss_download_repack_proper_episodes":true,
                    "rss_max_articles_per_feed":50,
                    "rss_processing_enabled":true,
                    "rss_refresh_interval":30,
                    "rss_smart_episode_filters":"s(\\d+)e(\\d+)\n(\\d+)x(\\d+)\n(\\d{4}[.\\-]\\d{1,2}[.\\-]\\d{1,2})",
                    "save_path": "/home/user/Downloads/",
                    "save_path_changed_tmm_enabled": false,
                    "save_resume_data_interval": 60,
                    "scan_dirs":
                    {
                        "/home/user/Downloads/incoming/games": 0,
                        "/home/user/Downloads/incoming/movies": 1
                    },
                    "schedule_from_hour": 8,
                    "schedule_from_min": 0,
                    "schedule_to_hour": 20,
                    "schedule_to_min": 0,
                    "scheduler_days": 0,
                    "scheduler_enabled": false,
                    "send_buffer_low_watermark": 10,
                    "send_buffer_watermark": 500,
                    "send_buffer_watermark_factor": 50,
                    "slow_torrent_dl_rate_threshold": 2,
                    "slow_torrent_inactive_timer": 60,
                    "slow_torrent_ul_rate_threshold": 2,
                    "socket_backlog_size": 30,
                    "start_paused_enabled": false,
                    "stop_tracker_timeout": 1,
                    "temp_path": "/home/user/Downloads/temp",
                    "temp_path_enabled": false,
                    "torrent_changed_tmm_enabled": true,
                    "up_limit": 0,
                    "upload_choking_algorithm": 1,
                    "upload_slots_behavior": 0,
                    "upnp": true,
                    "use_https": false,
                    "utp_tcp_mixed_mode": 0,
                    "web_ui_address": "*",
                    "web_ui_ban_duration": 3600,
                    "web_ui_clickjacking_protection_enabled": true,
                    "web_ui_csrf_protection_enabled": true,
                    "web_ui_custom_http_headers": "",
                    "web_ui_domain_list": "*",
                    "web_ui_host_header_validation_enabled": true,
                    "web_ui_https_cert_path": "",
                    "web_ui_https_key_path": "",
                    "web_ui_max_auth_fail_count": 5,
                    "web_ui_port": 8080,
                    "web_ui_secure_cookie_enabled": true,
                    "web_ui_session_timeout": 3600,
                    "web_ui_upnp": false,
                    "web_ui_use_custom_http_headers_enabled": false,
                    "web_ui_username": "admin"
                  }
      security:
        - sid: []
...
  /torrents/info:
    get:
      tags:
        - torrents
      summary: Get torrent list
      externalDocs:
        description: WebUI-API on qBittorrent wiki
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-list'
      operationId: torrentsInfoGet
      parameters:
        - in: query
          name: filter
          schema:
            type: string
            enum:
              - all
              - downloading
              - seeding
              - completed
              - paused
              - active
              - inactive
              - resumed
              - stalled
              - stalled_uploading
              - stalled_downloading
              - errored
          description: Filter torrent list by state.
        - in: query
          name: category
          schema:
            type: string
          description: Get torrents with the given category.
        - in: query
          name: tag
          schema:
            type: string
          description: Get torrents with the given tag.
        - in: query
          name: sort
          schema:
            type: string
          description: Sort torrents by given key.
        - in: query
          name: reverse
          schema:
            type: boolean
            default: false
          description: Enable reverse sorting.
        - in: query
          name: limit
          schema:
            type: integer
            format: int64
          description: Limit the number of torrents returned.
        - in: query
          name: offset
          schema:
            type: integer
            format: int64
          description: Set offset.
        - in: query
          name: hashes
          schema:
            type: array
            items:
              type: string
          description: Filter by hashes. Each hash should be provided as a separate parameter, e.g., hashes=value1&hashes=value2
      responses:
        '200':
          description: 'All scenarios- see JSON below'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TorrentInfo'
      security:
        - sid: []
    post:
      tags:
        - torrents
      summary: Get torrent list
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-torrent-list'
        description: WebUI-API on qBittorrent wiki
      operationId: torrentsInfoPost
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                filter:
                  type: string
                  enum:
                    - all
                    - downloading
                    - seeding
                    - completed
                    - paused
                    - active
                    - inactive
                    - resumed
                    - stalled
                    - stalled_uploading
                    - stalled_downloading
                    - errored
                  description: 'Filter torrent list by state. Allowed state filters: `all`, `downloading`, `seeding`, `completed`, `paused`, `active`, `inactive`, `resumed`, `stalled`, `stalled_uploading`, `stalled_downloading`, `errored`'
                category:
                  type: string
                  description: 'Get torrents with the given category (empty string means "without category"; no "category" parameter means "any category" <- broken until [#11748](https://github.com/qbittorrent/qBittorrent/issues/11748) is resolved). Remember to URL-encode the category name. For example, `My category` becomes `My%20category`'
                tag:
                  type: string
                  description: 'Get torrents with the given tag (empty string means "without tag"; no "tag" parameter means "any tag". Remember to URL-encode the category name. For example, `My tag` becomes `My%20tag`'
                sort:
                  type: string
                  description: Sort torrents by given key. They can be sorted using any field of the response's JSON array (which are documented below) as the sort key.
                reverse:
                  type: boolean
                  description: Enable reverse sorting. Defaults to `false`
                  default: false
                limit:
                  type: integer
                  format: int64
                  description: Limit the number of torrents returned
                offset:
                  type: integer
                  format: int64
                  description: 'Set offset (if less than 0, offset from end)'
                hashes:
                  type: array
                  items:
                    type: string
                  description: Filter by hashes. Can contain multiple hashes separated by `|`
            encoding:
              hashes:
                style: pipeDelimited
                explode: false
            example: /api/v2/torrents/info?filter=downloading&category=sample%20category&sort=ratio
        required: true
      responses:
        '200':
          description: All scenarios- see JSON below
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TorrentInfo'
                description: The response is a JSON array with the following fields
                example: |
                  [
                    {
                      "dlspeed":9681262,
                      "eta":87,
                      "f_l_piece_prio":false,
                      "force_start":false,
                      "hash":"8c212779b4abde7c6bc608063a0d008b7e40ce32",
                      "category":"",
                      "tags": "",
                      "name":"debian-8.1.0-amd64-CD-1.iso",
                      "num_complete":-1,
                      "num_incomplete":-1,
                      "num_leechs":2,
                      "num_seeds":54,
                      "priority":1,
                      "progress":0.16108787059783936,
                      "ratio":0,
                      "seq_dl":false,
                      "size":657457152,
                      "state":"downloading",
                      "super_seeding":false,
                      "upspeed":0
                    }
                  ]
      security:
        - sid: []
...
  /torrents/categories:
    get:
      tags:
        - torrents
      summary: Get all categories
      externalDocs:
        url: 'https://github.com/qbittorrent/qBittorrent/wiki/WebUI-API-(qBittorrent-4.1)#get-all-categories'
        description: WebUI-API on qBittorrent wiki
      operationId: torrentsCategoriesGet
      responses:
        '200':
          description: All scenarios
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  $ref: '#/components/schemas/TorrentsCategory'
                example: |
                  {
                      "Video": {
                          "name": "Video",
                          "savePath": "/home/user/torrents/video/"
                      },
                      "eBooks": {
                          "name": "eBooks",
                          "savePath": "/home/user/torrents/eBooks/"
                      }
                  }
      security:
        - sid: []
...
components:
  schemas:
...
    TorrentInfo:
      type: object
      properties:
        added_on:
          type: integer
          format: int64
          description: Time (Unix Epoch) when the torrent was added to the client
        amount_left:
          type: integer
          format: int64
          description: Amount of data left to download (bytes)
        auto_tmm:
          type: boolean
          description: Whether this torrent is managed by Automatic Torrent Management
        availability:
          type: number
          format: float
          description: Percentage of file pieces currently available
        category:
          type: string
          description: Category of the torrent
        completed:
          type: integer
          format: int64
          description: Amount of transfer data completed (bytes)
        completion_on:
          type: integer
          format: int64
          description: Time (Unix Epoch) when the torrent completed
        content_path:
          type: string
          description: 'Absolute path of torrent content (root path for multifile torrents, absolute file path for singlefile torrents)'
        dl_limit:
          type: integer
          format: int64
          description: Torrent download speed limit (bytes/s). `-1` if ulimited.
        dlspeed:
          type: integer
          format: int64
          description: Torrent download speed (bytes/s)
        downloaded:
          type: integer
          format: int64
          description: Amount of data downloaded
        downloaded_session:
          type: integer
          format: int64
          description: Amount of data downloaded this session
        eta:
          type: integer
          format: int64
          description: Torrent ETA (seconds)
        f_l_piece_prio:
          type: boolean
          description: True if first last piece are prioritized
        force_start:
          type: boolean
          description: True if force start is enabled for this torrent
        hash:
          type: string
          description: Torrent hash
        last_activity:
          type: integer
          format: int64
          description: Last time (Unix Epoch) when a chunk was downloaded/uploaded
        magnet_uri:
          type: string
          description: Magnet URI corresponding to this torrent
        max_ratio:
          type: number
          format: float
          description: Maximum share ratio until torrent is stopped from seeding/uploading
        max_seeding_time:
          type: integer
          format: int64
          description: Maximum seeding time (seconds) until torrent is stopped from seeding
        name:
          type: string
          description: Torrent name
        num_complete:
          type: integer
          format: int64
          description: Number of seeds in the swarm
        num_incomplete:
          type: integer
          format: int64
          description: Number of leechers in the swarm
        num_leechs:
          type: integer
          format: int64
          description: Number of leechers connected to
        num_seeds:
          type: integer
          format: int64
          description: Number of seeds connected to
        priority:
          type: integer
          format: int64
          description: Torrent priority. Returns -1 if queuing is disabled or torrent is in seed mode
        progress:
          type: number
          format: float
          description: Torrent progress (percentage/100)
        ratio:
          type: number
          format: float
          description: 'Torrent share ratio. Max ratio value: 9999.'
        ratio_limit:
          type: number
          format: float
          description: TODO (what is different from max_ratio?)
        save_path:
          type: string
          description: Path where this torrent's data is stored
        seeding_time:
          type: integer
          format: int64
          description: Torrent elapsed time while complete (seconds)
        seeding_time_limit:
          type: integer
          format: int64
          description: 'TODO (what is different from `max_seeding_time`?) seeding_time_limit is a per torrent setting, when Automatic Torrent Management is disabled, furthermore then max_seeding_time is set to seeding_time_limit for this torrent. If Automatic Torrent Management is enabled, the value is -2. And if max_seeding_time is unset it have a default value -1.'
        seen_complete:
          type: integer
          format: int64
          description: Time (Unix Epoch) when this torrent was last seen complete
        seq_dl:
          type: boolean
          description: True if sequential download is enabled
        size:
          type: integer
          format: int64
          description: Total size (bytes) of files selected for download
        state:
          type: string
          enum:
            - error
            - missingFiles
            - uploading
            - pausedUP
            - queuedUP
            - stalledUP
            - checkingUP
            - forcedUP
            - allocating
            - downloading
            - metaDL
            - pausedDL
            - queuedDL
            - stalledDL
            - checkingDL
            - forcedDL
            - checkingResumeData
            - moving
            - unknown
          description: |
            Torrent state. See table here below for the possible values

            | Value              | Description                                                   |
            | ------------------ | ------------------------------------------------------------- |
            | error              | Some error occurred, applies to paused torrents               |
            | missingFiles       | Torrent data files is missing                                 |
            | uploading          | Torrent is being seeded and data is being transferred         |
            | pausedUP           | Torrent is paused and has finished downloading                |
            | queuedUP           | Queuing is enabled and torrent is queued for upload           |
            | stalledUP          | Torrent is being seeded, but no connection were made          |
            | checkingUP         | Torrent has finished downloading and is being checked         |
            | forcedUP           | Torrent is forced to uploading and ignore queue limit         |
            | allocating         | Torrent is allocating disk space for download                 |
            | downloading        | Torrent is being downloaded and data is being transferred     |
            | metaDL             | Torrent has just started downloading and is fetching metadata |
            | pausedDL           | Torrent is paused and has NOT finished downloading            |
            | queuedDL           | Queuing is enabled and torrent is queued for download         |
            | stalledDL          | Torrent is being downloaded, but no connection were made      |
            | checkingDL         | Same as checkingUP, but torrent has NOT finished downloading  |
            | forcedDL           | Torrent is forced to downloading to ignore queue limit        |
            | checkingResumeData | Checking resume data on qBt startup                           |
            | moving             | Torrent is moving to another location                         |
            | unknown            | Unknown status                                                |
        super_seeding:
          type: boolean
          description: True if super seeding is enabled
        tags:
          type: string
          description: Comma-concatenated tag list of the torrent
        time_active:
          type: integer
          format: int64
          description: Total active time (seconds)
        total_size:
          type: integer
          format: int64
          description: Total size (bytes) of all file in this torrent (including unselected ones)
        tracker:
          type: string
          description: The first tracker with working status. Returns empty string if no tracker is working.
        up_limit:
          type: integer
          format: int64
          description: Torrent upload speed limit (bytes/s). `-1` if ulimited.
        uploaded:
          type: integer
          format: int64
          description: Amount of data uploaded
        uploaded_session:
          type: integer
          format: int64
          description: Amount of data uploaded this session
        upspeed:
          type: integer
          format: int64
          description: Torrent upload speed (bytes/s)
...
    TorrentsCategory:
      type: object
      properties:
        name:
          type: string
        savePath:
          type: string
...
  securitySchemes:
    sid:
      type: apiKey
      name: SID
      in: cookie
...